labeled obsolete versions and replaced input files with actual LPG output
