+++ hisim/inputs/loadprofiles/electrical-load_2-smart-appliances/LPG_output/FlexibilityEventsStatistics.HH1.json
@@ -1,0 +1,4 @@
+Total Events:555
+Total devices:2
+Dishwashers / Bosch SMI68 M35 EU; 324,22927776628285 kWh
+Washing Machine / Bosch WFS 8000; 280,98183939172316 kWh
