--- hisim/inputs/loadprofiles/electrical-load_2-smart-appliances/data_raw/FlexibilityEventsStatistics.HH1.json
@@ -1,4 +1,0 @@
-Total Events:316
-Total devices:2
-Washing Machine AEG Öko Plus 1400; 114,70653952670988 kWh
-Dishwashers / Bosch SMI68 M35 EU; 132,1357117046242 kWh
