+++ hisim/inputs/loadprofiles/electrical-load_2-smart-appliances/readme.md
@@ -1,21 +1,1 @@
-# electrical load of a tool manufacturer, located in Germany
-
-**Source & License**:
-
- [![License: CC BY 4.0](https://img.shields.io/badge/License-CC%20BY%204.0-lightgrey.svg)](https://creativecommons.org/licenses/by/4.0/)
-
-**Summary**:
-
-Data contains electricity consumption profiles of a smart washing machine and dishwasher with restricted time windows.
-
-**Content**
-
-* **number of columns**: 1 tool manufacturer
-* **length**: 1 year
-* **time increment**: 15min
-
-| datetime | electrical load [W] |
-| :--- | :--- |
-| 2019-01-01 00:00:00+01:00 | 15680 |
-| 2019-01-01 00:15:00+01:00 | 15520 |
-| ... | ... |
+﻿# output of load profile generator if flexibility calculations are enabled in calculation settings (apart from version 10.8)
